+++ index.js
@@ -14,6 +14,19 @@ const lennyWeaponsURL = "https://leanny.github.io/data/Mush/latest/WeaponInfo_Ma
 
 const gearTypeLang = require("./lang/gearTypeLang.json");
 
+const ruleImg_turfwar = "https://splatoon3.ink/assets/regular.81d2e9e4.svg";
+const ruleImg_rainmaker = "https://splatoon3.ink/assets/hoko.e3dce940.svg";
+const ruleImg_clamblitz = "https://splatoon3.ink/assets/asari.83043125.svg";
+const ruleImg_splatzones = "https://file.strassburger.org/splatzones.svg";
+const ruleImg_towercontrol = "https://file.strassburger.org/towercontrol.svg";
+
+function getImageFromRuleId(ruleId) {
+    if (ruleId === "VnNSdWxlLTQ=") return ruleImg_clamblitz;
+    else if (ruleId === "VnNSdWxlLTI=") return ruleImg_towercontrol;
+    else if (ruleId === "VnNSdWxlLTE=") return ruleImg_splatzones;
+    else if (ruleId === "VnNSdWxlLTM=") return ruleImg_rainmaker;
+    else return ruleImg_turfwar;
+}
 
 // https://stackoverflow.com/a/73401564
 function RGBAToHexA(rgba, forceRemoveAlpha = false) {
@@ -207,7 +220,8 @@ class Client {
                             name: translation.stages[json.data.regularSchedules.nodes[0].regularMatchSetting.vsStages[1].id].name,
                                 image: json.data.regularSchedules.nodes[0].regularMatchSetting.vsStages[1].image.url
                             },
-                            rules: translation.rules[json.data.regularSchedules.nodes[0].regularMatchSetting.vsRule.id].name
+                            rules: translation.rules[json.data.regularSchedules.nodes[0].regularMatchSetting.vsRule.id].name,
+                            rulesImg: getImageFromRuleId(json.data.regularSchedules.nodes[0].regularMatchSetting.vsRule.id)
                         }
                     } else {
                         data.regular = null;
@@ -226,7 +240,8 @@ class Client {
                                     name: translation.stages[json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[0].vsStages[1].id].name,
                                     image: json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[0].vsStages[1].image.url
                                 },
-                                rules: translation.rules[json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[0].vsRule.id].name
+                                rules: translation.rules[json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[0].vsRule.id].name,
+                                rulesImg: getImageFromRuleId(json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[0].vsRule.id)
                             },
                             open: {
                                 start_time: json.data.bankaraSchedules.nodes[0].startTime,
@@ -239,7 +254,8 @@ class Client {
                                     name: translation.stages[json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[1].vsStages[1].id].name,
                                     image: json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[1].vsStages[1].image.url
                                 },
-                                rules: translation.rules[json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[1].vsRule.id].name
+                                rules: translation.rules[json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[1].vsRule.id].name,
+                                rulesImg: getImageFromRuleId(json.data.bankaraSchedules.nodes[0].bankaraMatchSettings[1].vsRule.id)
                             }
                         }
                     } else {
@@ -258,10 +274,42 @@ class Client {
                                 name: translation.stages[json.data.xSchedules.nodes[0].xMatchSetting.vsStages[1].id].name,
                                 image: json.data.xSchedules.nodes[0].xMatchSetting.vsStages[1].image.url
                             },
-                            rules: translation.rules[json.data.xSchedules.nodes[0].xMatchSetting.vsRule.id].name
+                            rules: translation.rules[json.data.xSchedules.nodes[0].xMatchSetting.vsRule.id].name,
+                            rulesImg: getImageFromRuleId(json.data.xSchedules.nodes[0].xMatchSetting.vsRule.id),
                         }
                     } else {
                         data.xbattle = null;
+                    }
+
+                    if (json.data.festSchedules.nodes[0].festMatchSetting) {
+                        data.festSchedule = {
+                            start_time: json.data.festSchedules.nodes[0].startTime,
+                            end_time: json.data.festSchedules.nodes[0].endTime,
+                            stage1: {
+                                name: translation.stages[json.data.festSchedules.nodes[0].festMatchSetting.vsStages[0].id].name,
+                                image: json.data.festSchedules.nodes[0].festMatchSetting.vsStages[0].image.url
+                            },
+                            stage2: {
+                                name: translation.stages[json.data.festSchedules.nodes[0].festMatchSetting.vsStages[1].id].name,
+                                image: json.data.festSchedules.nodes[0].festMatchSetting.vsStages[1].image.url
+                            },
+                            rules: translation.rules[json.data.festSchedules.nodes[0].festMatchSetting.vsRule.id].name,
+                            rulesImg: getImageFromRuleId(json.data.festSchedules.nodes[0].festMatchSetting.vsRule.id),
+                        }
+                    } else {
+                        data.festSchedule = null;
+                    }
+
+                    if (json.data.currentFest) {
+                        data.triColorStage = {
+                            start_time: json.data.currentFest.startTime,
+                            end_time: json.data.currentFest.endTime,
+                            name: this.translation.stages[json.data.currentFest.tricolorStage.id].name,
+                            image: json.data.currentFest.tricolorStage.image.url,
+                            rulesImg: "https://file.strassburger.org/tricolor.svg",
+                        }
+                    } else {
+                        data.triColorStage = null;
                     }
 
                     return callback(data);
@@ -277,7 +325,7 @@ class Client {
                 .then(res => res.json())
                 .then(json => {
                     let data = {};
-                    if (json.data.regularSchedules.nodes[0].regularMatchSetting) {
+                    if (json.data.regularSchedules.nodes[1].regularMatchSetting) {
                         data.regular = {
                             start_time: json.data.regularSchedules.nodes[1].startTime,
                             end_time: json.data.regularSchedules.nodes[1].endTime,
@@ -289,13 +337,14 @@ class Client {
                             name: this.translation.stages[json.data.regularSchedules.nodes[1].regularMatchSetting.vsStages[1].id].name,
                                 image: json.data.regularSchedules.nodes[1].regularMatchSetting.vsStages[1].image.url
                             },
-                            rules: this.translation.rules[json.data.regularSchedules.nodes[1].regularMatchSetting.vsRule.id].name
+                            rules: this.translation.rules[json.data.regularSchedules.nodes[1].regularMatchSetting.vsRule.id].name,
+                            rulesImg: getImageFromRuleId(json.data.regularSchedules.nodes[1].regularMatchSetting.vsRule.id)
                         }
                     } else {
                         data.regular = null;
                     }
 
-                    if (json.data.bankaraSchedules.nodes[0].bankaraMatchSettings) {
+                    if (json.data.bankaraSchedules.nodes[1].bankaraMatchSettings) {
                         data.ranked = {
                             series: {
                                 start_time: json.data.bankaraSchedules.nodes[1].startTime,
@@ -308,7 +357,8 @@ class Client {
                                     name: this.translation.stages[json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[0].vsStages[1].id].name,
                                     image: json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[0].vsStages[1].image.url
                                 },
-                                rules: this.translation.rules[json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[0].vsRule.id].name
+                                rules: this.translation.rules[json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[0].vsRule.id].name,
+                                rulesImg: getImageFromRuleId(json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[0].vsRule.id),
                             },
                             open: {
                                 start_time: json.data.bankaraSchedules.nodes[1].startTime,
@@ -321,14 +371,15 @@ class Client {
                                     name: this.translation.stages[json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[1].vsStages[1].id].name,
                                     image: json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[1].vsStages[1].image.url
                                 },
-                                rules: this.translation.rules[json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[1].vsRule.id].name
+                                rules: this.translation.rules[json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[1].vsRule.id].name,
+                                rulesImg: getImageFromRuleId(json.data.bankaraSchedules.nodes[1].bankaraMatchSettings[1].vsRule.id)
                             }
                         }
                     } else {
                         data.ranked = null;
                     }
 
-                    if (json.data.xSchedules.nodes[0].xMatchSetting) {
+                    if (json.data.xSchedules.nodes[1].xMatchSetting) {
                         data.xbattle = {
                             start_time: json.data.xSchedules.nodes[1].startTime,
                             end_time: json.data.xSchedules.nodes[1].endTime,
@@ -340,11 +391,44 @@ class Client {
                             name: this.translation.stages[json.data.xSchedules.nodes[1].xMatchSetting.vsStages[1].id].name,
                                 image: json.data.xSchedules.nodes[1].xMatchSetting.vsStages[1].image.url
                             },
-                            rules: this.translation.rules[json.data.xSchedules.nodes[1].xMatchSetting.vsRule.id].name
+                            rules: this.translation.rules[json.data.xSchedules.nodes[1].xMatchSetting.vsRule.id].name,
+                            rulesImg: getImageFromRuleId(json.data.xSchedules.nodes[1].xMatchSetting.vsRule.id),
                         }
                     } else {
                         data.xbattle = null;
                     }
+
+                    if (json.data.festSchedules.nodes[1].festMatchSetting) {
+                        data.festSchedule = {
+                            start_time: json.data.festSchedules.nodes[1].startTime,
+                            end_time: json.data.festSchedules.nodes[1].endTime,
+                            stage1: {
+                                name: this.translation.stages[json.data.festSchedules.nodes[1].festMatchSetting.vsStages[0].id].name,
+                                image: json.data.festSchedules.nodes[1].festMatchSetting.vsStages[0].image.url
+                            },
+                            stage2: {
+                            name: this.translation.stages[json.data.festSchedules.nodes[1].festMatchSetting.vsStages[1].id].name,
+                                image: json.data.festSchedules.nodes[1].festMatchSetting.vsStages[1].image.url
+                            },
+                            rules: this.translation.rules[json.data.festSchedules.nodes[1].festMatchSetting.vsRule.id].name,
+                            rulesImg: getImageFromRuleId(json.data.festSchedules.nodes[1].festMatchSetting.vsRule.id),
+                        }
+                    } else {
+                        data.festSchedule = null;
+                    }
+
+                    if (json.data.currentFest) {
+                        data.triColorStage = {
+                            start_time: json.data.currentFest.startTime,
+                            end_time: json.data.currentFest.endTime,
+                            name: this.translation.stages[json.data.currentFest.tricolorStage.id].name,
+                            image: json.data.currentFest.tricolorStage.image.url,
+                            rulesImg: "https://file.strassburger.org/tricolor.svg",
+                        }
+                    } else {
+                        data.triColorStage = null;
+                    }
+
                     return callback(data);
                 });
         })
@@ -365,6 +449,7 @@ class Client {
                             desc: this.translation.events[event.leagueMatchSetting.leagueMatchEvent.id].desc,
                             eventRule: this.translation.events[event.leagueMatchSetting.leagueMatchEvent.id].regulation,
                             gameRule: this.translation.rules[event.leagueMatchSetting.vsRule.id].name,
+                            gameRuleImg: getImageFromRuleId(event.leagueMatchSetting.vsRule.id),
                             stages: [],
                             timePeriods: [],
                         }
@@ -881,8 +966,6 @@ class Client {
                             }
                         })
                     })
-
-                    return callback(data);
                 });
         })
     }