+++ CHANGELOG.md
@@ -4,6 +4,19 @@ Here, you will find all notabled changes that have been made to this project.
 
 ---
 
+## 1.4.3 - 2023-07-16
+
+### Added
+- Added rules images to rotations
+- Added Splatfest schedules
+- Added Tricolor stage
+
+### Changed
+
+### Fixed
+
+---
+
 ## 1.4.2 - 2023-07-16
 
 ### Added